GPU: Fix possible shift truncation.
Also, add algorithm for std::max.

+++ GPU/Common/TextureShaderCommon.cpp
@@ -191,7 +191,7 @@ Draw2DPipeline *TextureShaderCache::GetDepalettizeShader(uint32_t clutMode, GETe
 	using namespace Draw;
 
 	// Generate an ID for depal shaders.
-	u64 id = (depthUpperBits << 32) | (clutMode & 0xFFFFFF) | (textureFormat << 24) | (bufferFormat << 28);
+	u64 id = ((u64)depthUpperBits << 32) | (clutMode & 0xFFFFFF) | (textureFormat << 24) | (bufferFormat << 28);
 
 	auto shader = depalCache_.find(id);
 	if (shader != depalCache_.end()) {
